+++ src/components/days-list.js
@@ -1,0 +1,8 @@
+const getDaysListMarkup = () => {
+  return (
+    `<ul class="trip-days">
+    </ul>`
+  );
+};
+
+export default getDaysListMarkup;
